+++ FLySMPS/diodebridge.h
@@ -14,12 +14,12 @@ public:
     DiodeBridge();
 
     double IDiodePeak(double ICBulckPeak, double ILoadMax);
-    double IDiodeRMS(double ILoadAVG, double FLine, double ConductTime);
+    double IDiodeRMS(double ILoadAVG, int FLine, double ConductTime);
     double IDiodeAVG(double ILoadAVG);
 
-    double IDiodeRMSTot(double ILoadAVG, double FLine, double ConductTime);
+    double IDiodeRMSTot(double ILoadAVG, int FLine, double ConductTime);
 
-    double ILoadAVG(double IDiodePeak, double FLine, double ConductTime);
+    double ILoadAVG(double IDiodePeak, int FLine, double ConductTime);
     
     double DiodeCurrentSlope(double IDiodePeak, double ILoad, double TotCharTime);
     double DiodeConductTime(double IDiodePeak, double DiodeCurrentSlope);
